+++ docs/source/reference/trackers.rst
@@ -1,0 +1,5 @@
+Result Trackers
+===============
+.. automodapi:: pykeen.trackers
+    :no-heading:
+    :headings: --
